+++ LookupAnything/Framework/ModConfigKeys.cs
@@ -15,11 +15,17 @@ namespace Pathoschild.Stardew.LookupAnything.Framework
         /// <summary>The keys which toggle the search UI.</summary>
         public KeybindList ToggleSearch { get; set; } = KeybindList.Parse($"{SButton.LeftShift} + {SButton.F1}");
 
-        /// <summary>The keys which scroll up long content.</summary>
+        /// <summary>The keys which scroll up long content by a few lines.</summary>
         public KeybindList ScrollUp { get; set; } = new(SButton.Up);
 
-        /// <summary>The keys which scroll down long content.</summary>
+        /// <summary>The keys which scroll down long content by a few lines.</summary>
         public KeybindList ScrollDown { get; set; } = new(SButton.Down);
+
+        /// <summary>The keys which scroll up long content by a full page.</summary>
+        public KeybindList PageUp { get; set; } = new(SButton.PageUp);
+
+        /// <summary>The keys which scroll down long content by a full page.</summary>
+        public KeybindList PageDown { get; set; } = new(SButton.PageDown);
 
         /// <summary>The keys which toggle the display of debug information.</summary>
         public KeybindList ToggleDebug { get; set; } = new();